AI chatbot /roadies-backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
woocommerce
httpx
pydantic